requirements.txt
pdfplumber
langdetect
requests
python-dotenv
